Move update methods into RenderStateExtensions
